% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_overwrite.R
\name{check_file_overwrite}
\alias{check_file_overwrite}
\title{Check if a file exists and remove file or error}
\usage{
check_file_overwrite(
  filename = NULL,
  path = NULL,
  overwrite = TRUE,
  quiet = FALSE,
  ask = TRUE,
  call = caller_env()
)
}
\arguments{
\item{filename}{File name, Default: \code{NULL}}

\item{path}{File path, Default: \code{NULL}}

\item{overwrite}{If \code{TRUE}, remove a file with the same name and path}

\item{quiet}{If \code{TRUE}, suppress informational messages, Default: \code{FALSE}}

\item{ask}{If \code{TRUE}, overwrite is \code{FALSE}, and session is interactive, ask
if user wants to overwrite the file. Default: \code{TRUE}}

\item{call}{Passed to \code{\link[=cli_abort]{cli_abort()}}, Default: \code{\link[=caller_env]{caller_env()}}}
}
\description{
Check if a file exists and remove file or error
}
