% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_exif_tags}
\alias{default_exif_tags}
\title{Default EXIF, XMP-dc, and IPTC tags}
\format{
A length 26 character vector.
}
\usage{
default_exif_tags
}
\description{
A vector of default EXIF, XMP-dc, and IPTC tags (including wildcard EXIF tag
"GPS") used by \code{\link[=read_exif]{read_exif()}}.
}
\keyword{datasets}
