% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_exif_xwalk}
\alias{default_exif_xwalk}
\title{EXIF data column name crosswalk}
\format{
A length 29 named character vector.
}
\usage{
default_exif_xwalk
}
\description{
A named vector with a crosswalk of default column names returned by
\code{\link[exiftoolr:exif_read]{exiftoolr::exif_read()}} and the replacement values used as new names.
}
\keyword{datasets}
