% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_exif.R
\name{fmt_exif_direction}
\alias{fmt_exif_direction}
\title{Format a data.frame column as a cardinal direction in degrees and wind direction}
\usage{
fmt_exif_direction(data, winds = 8, .after = "img_direction")
}
\arguments{
\item{data}{A data.frame or character vector to format with
\code{\link[isstatic:as_cardinal_bearing]{isstatic::as_cardinal_bearing()}}. If data is a data.frame, the photo
direction is assumed to be in a column named "img_direction" and the new
cardinal bearing is added to a column named "img_cardinal_wind".}

\item{winds}{Number of winds to use for results (4, 8, or 16).}

\item{.after}{Column name passed to .after parameter of \code{\link[dplyr:mutate]{dplyr::mutate()}}.}
}
\description{
Format a data.frame column as a cardinal direction in degrees and wind direction
}
