% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_fileext.R
\name{get_path_fileext}
\alias{get_path_fileext}
\alias{list_path_fileext}
\title{Get file types or extensions for files at a path}
\usage{
get_path_fileext(
  path,
  fileext = NULL,
  n = 1,
  quiet = FALSE,
  call = caller_env()
)

list_path_fileext(path, allow_null = FALSE, call = caller_env(), ...)
}
\arguments{
\item{path}{A single directory or file path. The directory or file must
exist.}

\item{fileext}{If not \code{NULL}, function returns file type as is.}

\item{n}{Max number of unique file types to return. Returns warning and n
most common file types if path has more than n unique file types.}

\item{quiet}{If \code{TRUE}, suppress informational messages.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
If fileext is provided, \code{\link[=get_path_fileext]{get_path_fileext()}} will pass the file extension
forward without checking it. \code{\link[=list_path_fileext]{list_path_fileext()}} is a more basic function
that list file extensions at a path directory.
}
