% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pkg_data.R
\name{list_pkg_data}
\alias{list_pkg_data}
\alias{list_pkg_datasets}
\alias{list_pkg_extdata}
\alias{list_pkg_cachedata}
\title{List package datasets, extdata files, and files in package cache}
\usage{
list_pkg_data(pkg = NULL, version = NULL, call = caller_env())

list_pkg_datasets(pkg, lib.loc = NULL)

list_pkg_extdata(pkg, dir = "extdata", full.names = TRUE, recursive = TRUE)

list_pkg_cachedata(pkg, full.names = TRUE, recursive = TRUE)
}
\arguments{
\item{pkg}{The package names. Can include version requirements,
e.g. \code{"pkg (>= 1.0.0)"}.}

\item{version}{Minimum versions for \code{pkg}. If supplied, must be the
same length as \code{pkg}. \code{NA} elements stand for any versions.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{lib.loc}{a character vector with path names of \R libraries.
    See \sQuote{Details} for the meaning of the default value of \code{NULL}.}

\item{dir}{One or more directory names to pass as the first argument of
\code{\link[=system.file]{system.file()}}. Defaults to "extdir".}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}
}
\description{
\code{\link[=list_pkg_data]{list_pkg_data()}} calls three helper functions and returns a data.frame.
\code{\link[=list_pkg_datasets]{list_pkg_datasets()}} lists package datasets (using the library path as the
path), \code{\link[=list_pkg_extdata]{list_pkg_extdata()}} lists extdata files for a package (using basename
without a file extension as the item name), \code{\link[=list_pkg_cachedata]{list_pkg_cachedata()}} lists data
in the package cache folder returned by \code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}.
}
