% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make file name and path with optional label, prefix, or postfix}
\usage{
make_filename(
  name = NULL,
  label = NULL,
  fileext = NULL,
  filename = NULL,
  path = NULL,
  prefix = NULL,
  postfix = NULL,
  pad = NULL,
  width = NULL,
  cache = FALSE,
  appname = NULL,
  create = TRUE,
  increment = NULL
)
}
\arguments{
\item{name}{Name to make file name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}, e.g. "Residential zoning map" becomes
"residential_zoning_map". If the name includes a file extension it is
assumed that the filename has been provided as the name parameter.}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}. The label is designed to identify the area
or other shared characteristics across multiple data files, maps, or plots.
label is ignored if name is NULL or if name includes a file extension.}

\item{fileext}{File type or extension. Optional if filename or path include a
file extension.}

\item{filename}{File name; if filename is \code{NULL} and path does not include a
file extension, name and file extension are both required.}

\item{path}{Path to file or data directory. Optional. If path includes a file
extension and filename and fileext are both \code{NULL}, the filename and
extension included with path will be used instead. If multiple file
extensions are provided to filename, path, or fileext, \code{make_filename()}
will abort.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{width}{Minimum width of padded strings.}

\item{cache}{If \code{TRUE}, path is set to the package cache directory using
\code{\link[=get_data_dir]{get_data_dir()}}; defaults to \code{FALSE}.}

\item{appname}{Passed to \code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}}

\item{create}{If \code{FALSE} and path does not exist, return path with a warning.
If \code{TRUE} and \code{\link[rlang:is_interactive]{rlang::is_interactive()}} is \code{TRUE}, ask user if directory
should be created. If the session not interactive and create is \code{TRUE}, a
new directory will be created.}

\item{increment}{If \code{TRUE}, increment digits in string by 1. If numeric,
increment digits in string by value. If \code{NULL}, 0, or if no digits are
present in string, return string as is.}
}
\description{
A helper function to create consistent file names for plots created with data
files.
}
\examples{

make_filename(
  filename = "image.jpeg"
)

make_filename(
  name = "plot",
  label = "Group a",
  fileext = "png"
)

make_filename(
  name = "plot",
  prefix = "date",
  fileext = "png"
)

make_filename(
  name = "map_1",
  increment = TRUE,
  fileext = "geojson"
)

}
\concept{read_write}
