% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_affix.R
\name{str_affix}
\alias{str_affix}
\alias{str_prefix}
\title{Apply a prefix or postfix to a string}
\usage{
str_affix(
  string = NULL,
  prefix = NULL,
  postfix = NULL,
  sep = "_",
  pad = NULL,
  width = NULL,
  use_clean_names = TRUE,
  case = "snake",
  replace = c(`'` = "", `"` = "", `\%` = "_pct_", `#` = "_num_"),
  use_make_names = TRUE,
  ...
)

str_prefix(
  string = NULL,
  prefix = NULL,
  sep = "_",
  is_postfix = FALSE,
  date.format = "\%F",
  time.format = "\%Y-\%m-\%d_\%I-\%M-\%S_\%p",
  use_clean_names = TRUE,
  case = "snake",
  replace = c(`'` = "", `"` = "", `\%` = "_pct_", `#` = "_num_"),
  use_make_names = TRUE,
  ...
)
}
\arguments{
\item{string}{A single string that the attach prefix or postfix is added to.}

\item{prefix}{Character string or character vector to add to string parameter
as a prefix.}

\item{postfix}{Character string or character vector to add to string
parameter as a postfix.}

\item{sep}{Separator character passed as the collapse parameter of \code{\link[=paste]{paste()}}.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{width}{Minimum width of padded strings.}

\item{use_clean_names}{If \code{TRUE}, prefix, postfix, and string are all
converted to snake case with \code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}.}

\item{case}{The desired target case (default is \code{"snake"}) will be
passed to \code{snakecase::to_any_case()} with the exception of "old_janitor",
which exists only to support legacy code (it preserves the behavior of
\code{clean_names()} prior to addition of the "case" argument (janitor
versions <= 0.3.1).  "old_janitor" is not intended for new code. See
\code{\link[snakecase]{to_any_case}} for a wide variety of supported cases,
including "sentence" and "title" case.}

\item{replace}{A named character vector where the name is replaced by the
value.}

\item{use_make_names}{Should \code{make.names()} be applied to ensure that the
output is usable as a name without quoting?  (Avoiding \code{make.names()}
ensures that the output is locale-independent but quoting may be required.)}

\item{...}{Additional parameters passed to janitor::make_clean_names() if
use_clean_names is \code{TRUE}.}

\item{is_postfix}{If \code{TRUE}, use the prefix string as a postfix; defaults to
\code{FALSE}.}

\item{date.format, time.format}{Date or time format. Only used by \link{str_prefix}
if prefix is "date" or "time" and not currently accessible when using
\code{\link[=str_affix]{str_affix()}} or \code{\link[=make_filename]{make_filename()}}.}
}
\description{
Prefix and postfix can include more than one value that are added in the same
order provided. For \code{\link[=str_affix]{str_affix()}}, the string must be a single character
string.
}
\details{
\itemize{
\item \code{\link[=str_affix]{str_affix()}}: Add a label, prefix, and postfix to string
\item \code{\link[=str_prefix]{str_prefix()}}: Add a prefix or a postfix to a string
}
}
