% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_digits.R
\name{str_digits}
\alias{str_digits}
\alias{str_pad_digits}
\alias{str_extract_digits}
\alias{str_replace_digits}
\alias{str_increment_digits}
\title{Modify digits within strings}
\usage{
str_pad_digits(string, pad = "0", side = "left", width = NULL)

str_extract_digits(string, pattern = "[0-9]+", side = NULL)

str_replace_digits(string, replacement, pad = "0", side = "left", width = NULL)

str_increment_digits(string, increment = TRUE, ...)
}
\arguments{
\item{string}{Input vector.
Either a character vector, or something coercible to one.}

\item{pad}{Single padding character added to digits in string; defaults to
"0"}

\item{side}{Side on which padding character is added (left, right or both).}

\item{width}{Minimum width of padded strings.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression,
as described in \link[base:regex]{base::regex}.
Control options with \code{\link[stringstatic:regex]{regex()}}.

Match a fixed string (i.e. by comparing only bytes), using \code{\link[stringstatic:fixed]{fixed()}}.
This is fast, but approximate.}

\item{replacement}{A character vector of replacements.
Should be either length one, or the same length as \code{string} or \code{pattern}.
References of the form \verb{\\1}, \verb{\\2}, etc. will be replaced with the contents
of the respective matched group (created by \verb{()}).

To replace the complete string with \code{NA},
use \code{replacement = NA_character_}.

Using a function for \code{replacement} is not yet supported.}

\item{increment}{If \code{TRUE}, increment digits in string by 1. If numeric,
increment digits in string by value. If \code{NULL}, 0, or if no digits are
present in string, return string as is.}

\item{...}{Passed to \code{\link[=str_replace_digits]{str_replace_digits()}}}
}
\description{
\itemize{
\item \code{\link[=str_replace_digits]{str_replace_digits()}}: Replace digits with a string optionally
incrementing the digits
\item \code{\link[=str_pad_digits]{str_pad_digits()}}: Pad a string with digits
\item \code{\link[=str_extract_digits]{str_extract_digits()}}: Extract digits from a string
}
}
