% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_fileext.R
\name{str_fileext}
\alias{str_fileext}
\alias{str_add_fileext}
\alias{str_remove_fileext}
\alias{str_extract_fileext}
\title{Modify strings to help make consistent file names}
\usage{
str_add_fileext(string, fileext = NULL)

str_remove_fileext(string, fileext = NULL)

str_extract_fileext(string, fileext = NULL, tocase = tolower)
}
\arguments{
\item{string}{Character vector}

\item{fileext}{File extension string}

\item{tocase}{Function to use with \link{str_extract_fileext} for case conversion
of returned file extension Defaults to \code{\link[=tolower]{tolower()}}.}
}
\description{
Functions include:
}
\details{
\itemize{
\item \code{\link[=str_add_fileext]{str_add_fileext()}}: Add file type to string
\item \code{\link[=str_remove_fileext]{str_remove_fileext()}}: Remove file type from string
\item \code{\link[=str_extract_fileext]{str_extract_fileext()}}: Extract file type from string
\item \code{\link[=has_fileext]{has_fileext()}}: Does a string have a file extension (optionally
user-defined)?
}
}
