% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_overwrite.R
\name{check_file_overwrite}
\alias{check_file_overwrite}
\title{Check if a file exists and remove file or error}
\usage{
check_file_overwrite(
  filename = NULL,
  path = NULL,
  overwrite = TRUE,
  quiet = FALSE,
  ask = TRUE,
  call = caller_env()
)
}
\arguments{
\item{filename}{File name, Default: \code{NULL}. Optional if path is supplied.}

\item{path}{File path, Default: \code{NULL}. Optional if filename is supplied.}

\item{overwrite}{If \code{TRUE}, remove a file with the same name and path}

\item{quiet}{If \code{TRUE}, suppress informational messages, Default: \code{FALSE}}

\item{ask}{If \code{TRUE}, overwrite is \code{FALSE}, and session is interactive, ask
if user wants to overwrite the file. Default: \code{TRUE}}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
The filename or path must include a single file extension.
}
