% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_overwrite.R
\name{check_path_fileext}
\alias{check_path_fileext}
\title{Check if a file path has a file extension}
\usage{
check_path_fileext(
  path,
  fileext = NULL,
  message = "{.arg path} must include a file extension.",
  call = parent.frame()
)
}
\arguments{
\item{path}{File path to check. Required.}

\item{fileext}{Optional file extension. If \code{NULL}, path must have a file
extension. If a character string, path must have a matching file extension.}

\item{message}{It is formatted via a call to \code{\link[cli:cli_bullets]{cli_bullets()}}.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\description{
Check if a file path has a file extension
}
