% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenamr-package.R
\docType{package}
\name{filenamr-package}
\alias{filenamr}
\alias{filenamr-package}
\title{filenamr: Make and Modify File Names and Metadata}
\description{
Work with filenames and paths and read and write file metadata.
}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
