% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_fileext.R
\name{get_path_fileext}
\alias{get_path_fileext}
\alias{list_path_fileext}
\title{Get file extensions for files at a path}
\usage{
get_path_fileext(
  path,
  fileext = NULL,
  n = 1,
  quiet = FALSE,
  call = caller_env()
)

list_path_fileext(path, allow_null = FALSE, call = caller_env(), ...)
}
\arguments{
\item{path}{A single directory or file path. The directory or file must
exist.}

\item{fileext}{If fileext is supplied, the function returns the file
extension as is. If \code{NULL} (default), one or more file extensions are
extracted from files at the path location.}

\item{n}{Max number of unique file types to return. Returns warning and n
most common file types if path has more than n unique file types.}

\item{quiet}{If \code{TRUE}, suppress informational messages.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{allow_null}{If \code{TRUE}, \code{\link[=list_path_fileext]{list_path_fileext()}} returns \code{NULL} if path is
\code{NULL} or if no files exist at the path location. If \code{FALSE} (default),
abort if either condition is met.}

\item{...}{Additional parameters passed by \code{\link[=list_path_fileext]{list_path_fileext()}} to
\code{\link[=list.files]{list.files()}}.}
}
\description{
If fileext is provided, \code{\link[=get_path_fileext]{get_path_fileext()}} will pass the file extension
forward without checking it. \code{\link[=list_path_fileext]{list_path_fileext()}} is a more basic function
that list file extensions at a path directory.
}
