% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif.R
\name{read_exif}
\alias{read_exif}
\alias{write_exif}
\title{Read EXIF metadata to create a simple feature object or write
EXIF metadata to image files}
\usage{
read_exif(path = NULL, fileext = NULL, tags = NULL, ...)

write_exif(
  path,
  fileext = NULL,
  title = NULL,
  author = NULL,
  credit = author,
  date = NULL,
  keywords = NULL,
  description = NULL,
  alt = NULL,
  args = NULL,
  overwrite = TRUE,
  append_keywords = FALSE,
  call = caller_env()
)
}
\arguments{
\item{path}{A path to folder or file.}

\item{fileext}{The file extension or file type; defaults to \code{NULL}.}

\item{tags}{Optional list of EXIF tags to read from files. Must include GPS
tags if to create an \code{sf} object based on the resulting data.frame object.}

\item{...}{Additional EXIF tags to pass to \code{\link[exiftoolr:exif_read]{exiftoolr::exif_read()}}}

\item{title}{Title to add to file metadata with exiftoolr, Default: \code{NULL}.}

\item{author}{Author to add to file metadata to the "Author" and
"XMP-dc:creator" tags. Default: \code{NULL}.}

\item{credit}{Credit to add to file metadata to the "IPTC:Credit" and
"XMP-dc:Credit" tags. Defaults to the same value as author.}

\item{date}{Date to add to file metadata with exiftoolr (not currently
working).  Defaults to \code{NULL}.}

\item{keywords}{Keyword(s) added to file metadata to "IPTC:Keywords" and
"XMP-dc:Subject" tags. Defaults to \code{NULL}.}

\item{description}{Description added to the "ImageDescription",
"IPTC:Caption-Abstract", and "XMP-dc:Description" tags.}

\item{alt}{Text to pass as alt text to the "IPTC:AltTextAccessibility" and
"iTXt" (PNG files only) tags. Defaults to \code{NULL}.}

\item{args}{Alternate arguments passed to \code{\link[exiftoolr:exif_call]{exiftoolr::exif_call()}}. Other tag
parameters are appended to args if they are not \code{NULL}.}

\item{overwrite}{If \code{TRUE}, overwrite any existing EXIF metadata present in
the provided fields; defaults to \code{TRUE}}

\item{append_keywords}{If \code{TRUE}, append keywords, if \code{FALSE}, replace
keywords in file metadata.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=read_exif]{read_exif()}} read EXIF data from folder of files. This function also assigns
a cardinal direction based on the direction metadata and recodes the
orientation metadata.

For \code{\link[=write_exif]{write_exif()}} the parameters are used to multiple tags with the same
values:
\itemize{
\item title: Title, IPTC:Headline, IPTC:ObjectName, XMP-dc:Title
\item description: ImageDescription, XMP-dc:Description, and
IPTC:Caption-Abstract
\item keywords: Keywords, IPTC:Keywords, XMP-dc:Subject
}
}
