% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif.R
\name{read_exif}
\alias{read_exif}
\alias{write_exif}
\title{Read EXIF metadata to create a simple feature object or write
EXIF metadata to image files}
\usage{
read_exif(
  path = NULL,
  fileext = NULL,
  bbox = NULL,
  tags = getOption("read_exif.tags", default = default_tags),
  geometry = FALSE,
  ...
)

write_exif(
  path,
  fileext = NULL,
  title = NULL,
  author = NULL,
  date = NULL,
  keywords = NULL,
  description = NULL,
  args = NULL,
  overwrite = TRUE,
  append_keywords = FALSE
)
}
\arguments{
\item{path}{A path to folder or file.}

\item{fileext}{The file extension or file type; defaults to \code{NULL}.}

\item{bbox}{Optional bounding box to crop returned file (excluding images
with location data outside the bounding box). If bbox is provided the
returned data will match the crs of the bbox.}

\item{tags}{Optional list of EXIF tags to read from files. Must include GPS
tags to create an \code{sf} object.}

\item{geometry}{If \code{TRUE}, convert the data.frame with coordinates to a sf
object using \code{\link[sf:st_as_sf]{sf::st_as_sf()}}.}

\item{...}{Additional EXIF tags to pass to \link[exiftoolr:exif_read]{exiftoolr::exif_read}}

\item{title}{Title to add to file metadata with exiftoolr, Default: \code{NULL}.}

\item{author}{Author to add to file metadata with exiftoolr, Default: \code{NULL}.}

\item{date}{Date to add to file metadata with exiftoolr (not currently
working), Default: \code{NULL}.}

\item{keywords}{Keyword(s) added to file metadata with with exiftoolr,
Default: \code{NULL}.}

\item{description}{Description added to file metadata.}

\item{args}{Alternate arguments passed to \code{\link[exiftoolr:exif_call]{exiftoolr::exif_call()}}. If args
is not \code{NULL}, title, author, date, and keywords are ignored; defaults to
\code{NULL}.}

\item{overwrite}{If \code{TRUE}, overwrite any existing EXIF metadata present in
the provided fields; defaults to \code{TRUE}}

\item{append_keywords}{If \code{TRUE}, append keywords, if \code{FALSE}, replace
keywords in file metadata.}
}
\description{
\code{\link[=read_exif]{read_exif()}} read EXIF data from folder of files. This function also assigns
a cardinal direction based on the direction metadata and recodes the
orientation metadata.

For \code{\link[=write_exif]{write_exif()}} the parameters are used to multiple tags with the same
values:
\itemize{
\item title: Title, IPTC:Headline, IPTC:ObjectName, XMP-dc:Title
\item description: ImageDescription, XMP-dc:Description, and
IPTC:Caption-Abstract
\item keywords: Keywords, IPTC:Keywords, XMP-dc:Subject
}
}
