% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_file_path.R
\name{set_file_path}
\alias{set_file_path}
\title{Set a file path and validate path file extension}
\usage{
set_file_path(
  filename = NULL,
  path = NULL,
  fileext = NULL,
  allow_null = FALSE,
  call = caller_env()
)
}
\arguments{
\item{filename}{File name. Optional if path is supplied.}

\item{path}{File path. Optional if filename is supplied.}

\item{fileext}{File extension. If the path supplied using filename and path
\emph{does not} end with a file extension, fileext is used
as the file extension for the returned path. If the path, \emph{does} end with a
file extension, and fileext is used to validate the supplied.}

\item{allow_null}{If \code{TRUE}, return \code{NULL} if filename and path are \code{NULL}.
If \code{FALSE}, error if filename and path are both \code{NULL}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
This is a flexible wrapper for \code{\link[isstatic:file_path]{isstatic::file_path()}} that allows you to
provide a file path as a filename or path parameter. If the path contains a
file extension and the fileext parameter is provided, the function aborts if
the two file extensions do not match.
}
